Update viewfeed.js to allow control-clicks to open in background tabs
In tt-rss, control-clicks of article titles (collapsed or expended) do not seem to work as intended by the settings of my Web browser, namely to open in a background tab (instead they are opened in a foreground tab).

To solve this problem while being cross-browser compatible, I suggest using native clicks instead of the JavaScript function 'openArticleInNewWindow()' when clicking on a [A] element while the control-key is pressed.

Tested to work fine with Firefox 23, Chrome 30, IE 10 - IE 9.